requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Database
asyncpg>=0.29.0
redis>=5.0.1
sqlalchemy>=2.0.23

# MCP and Archon integration
mcp>=1.13.0
archon>=0.1.0

# Testing and HTTP client
aiohttp>=3.9.0

# Supabase integration
supabase>=2.18.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
requests>=2.31.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
